requirements.txt
Django==3.2.16
aiofiles==23.2.1
aiogram==2.23.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
certifi==2023.11.17
charset-normalizer==3.3.2
coloredlogs==15.0.1
exceptiongroup==1.2.0
executor==23.2
fasteners==0.19
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
humanfriendly==10.0
idna==3.6
magic-filter==1.0.12
multidict==6.0.4
property-manager==3.0
pydantic==2.5.3
pydantic_core==2.14.6
pyreadline3==3.4.1
python-dotenv==1.0.0
python-telegram-bot==20.7
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
telegram==0.0.1
typing_extensions==4.9.0
verboselogs==1.7
yarl==1.9.4